style: 优化 vxe-table 类型组织结构，将 viewed-row 功能域相关类型进行模块化拆分，由全局类型声明调整为功能目录内聚管理，避免公共 types 文件持续膨胀。
