+++ packages/effects/plugins/src/vxe-table/viewed-row/index.ts
@@ -1,0 +1,8 @@
+export type {
+  ViewedRowOptions,
+  ViewedRowPersistOptions,
+  ViewedRowStorageAdapter,
+} from './types';
+
+export { applyViewedRowOptions, useViewedRow } from './use-viewed-row';
+export type { ViewedRowHelper } from './use-viewed-row';
